refactor(archtest): Auth/Identity 核心 funnel 族迁 module-path-agnostic CellRule + 缩 baseline [#1632]
M3 PR-2: migrate the 10 Auth/Identity 核心 funnel archtest rule files to be
module-path-agnostic, following the #1084 PR-1 recipe (errcode/panic precedent).

For each rule the detector logic moves from `_test.go` into an importable
non-test `Check*(t, ConfigForExternalCell) []Diagnostic` function; platform
symbol paths derive from `external.go::PlatformModulePath` (no bare
"github.com/ghbvf/gocell" literal); the thinned `_test.go` dogfoods the same
Check (single source) and keeps its RED / anti-vacuity / negative-probe
self-checks. The 10 baseline lines are deleted from
`testdata/module_path_funnel.baseline` (ratchet-down, 64 → 54; was 66 → 56 pre-rebase onto develop which dropped 2 scaffold lines).

Migrated rules:
- CTXKEYS-PRINCIPAL-WRITE-CALLER-01      (ctxkeys_principal_write_caller)
- SAFEID-WIREMESSAGE-USAGE-01 / -UPSTREAM-FUNNEL-HARD-01 (safeid_funnel)
- FENCE-TOKEN-MINT-FUNNEL-01             (fence_token_mint_funnel)
- BCRYPT-COST-FUNNEL-01                  (bcrypt_cost_funnel)
- SVCTOKEN-CALLER-CELL-REQUIRED-01       (svctoken_caller_cell)
- NO-DELETED-AUTH-SYMBOLS-01             (no_deleted_auth_symbols)
- SEED-ROLE-IFACE-01                     (seed_role_iface)
- AUTH-AUTHTEST-BOUNDARY-01              (auth_authtest_boundary)
- AUTH-KEYSTEST-IMPORT-BOUNDARY-01       (auth_keystest_boundary)
- AUTH-PLAN-01/02/03/04                  (auth_plan)

Decisions:
- register=none: all 10 are gocell-internal-layout funnels whose allowlists are
  gocell-hardcoded with no ConfigForExternalCell consumer-extension, so they are
  importable + module-path-agnostic but stay OUT of StandardCellRules
  (external.go untouched); rationale self-documented in each new .go godoc.
- behavior 零变 incl. AI-robust rating: each rule's grade godoc moved intact;
  dogfood + RED self-checks prove the scanners are unchanged. The funnel itself
  advances toward its terminal Hard (empty-baseline) state by -10; it does NOT
  attempt the downstream-residual closure tracked separately at #1304.
- complexity: archtest scanners moved into production .go trip gocognit/cyclop
  (`.golangci.yml` excludes only _test.go); refactored into per-form helpers per
  the errcode/panic precedent. safeid_funnel keeps its author R2-approved
  nolint waivers (genuinely-linear type-resolution sequence).

Two archtests stay RED — PRE-EXISTING on origin/develop, unrelated to this
module-path migration, preserved byte-identically (not regressions):
- TestCtxkeysPrincipalWriteCaller01: cells/configcore/configcoretest/fakes.go:33
  calls ctxkeys.WithTenantID outside the allowlist (landed via tenancy #1479).
- TestSVCTOKEN_..._BuildTaggedFilesScanned_Wave5_RED: examples/ssobff (now a
  satellite module) skipped by tag scan — known #1590.

Verify:
- go build ./... ; go build -tags=integration ./...
- go test ./tools/archtest/ (all rule tests green except the 2 preserved reds)
- golangci-lint run ./tools/archtest/... → 0 issues

Refs: #1632 (M3 PR-2 of #1302); recipe #1084; Hard-ization #1304
