+++ tools/archtest/bcrypt_cost_funnel_test.go
@@ -1,68 +1,10 @@
 // INVARIANT: BCRYPT-COST-FUNNEL-01
-//
-// # BCRYPT-COST-FUNNEL-01
-//
-// All accesscore password hashing routes through the single
-// cells/accesscore/internal/credential.Hasher, whose bcrypt cost is chosen by
-// which constructor minted it:
-//
-//   - credential.NewProductionHasher()  — hardwired credential.ProductionCost
-//     (=12). No cost parameter exists, so the production path is structurally
-//     incapable of expressing a weaker cost.
-//   - credential.NewTestHasher(cost)    — the low-cost test door; cuts the
-//     ~1.5s/hash (bcrypt cost 12 under -race) that made cells/accesscore/slices/setup
-//     an 82s race-unit outlier and l2atomicity seedAdmin a per-test tax.
-//
-// Two rules close the funnel (per ai-robust.md §Hard 技术族目录):
-//
-//   - A1 "single sanctioned holder": bcrypt.GenerateFromPassword may appear
-//     only in credential/hasher.go. This is what makes the const-12 guarantee
-//     that the deleted domain.BcryptCost used to give survive injection — no
-//     other production file can hash at all, so none can pick a cost.
-//   - A2 "typed function choice": credential.NewTestHasher (the only cost-bearing
-//     door) may be called only from sanctioned test locations. Selecting the
-//     wrong semantics is selecting the wrong function name, not passing a wrong
-//     int — there is no "looks-right-but-isn't" gray zone of a literal that
-//     happens to equal a low cost.
-//
-// # AI-robust grade
-//
-// Downstream Hard (A1 callee-location + A2 caller-allowlist, both archtest-locked
-// by callsite identity). Upstream Medium: Go has no friend-package, so the
-// compiler cannot stop a NEW direct bcrypt.GenerateFromPassword call added
-// inside the credential package itself from bypassing Hasher.Hash; archtest
-// catches it in CI but the type system does not reject it. Upstream Hard-ization
-// is tracked by gh issue #901 (see §Funnel 双向锁评级).
-//
-// # Why pure-AST (not the typed façade)
-//
-// Both rules resolve a callsite by import-path + alias + selector name — no
-// receiver-type / interface-implementation / const-evaluation is needed. Per
-// ai-robust.md §载体决策原则, that is the "纯 AST 模式" route. Mirrors the sibling
-// PG-TESTCONTAINER-FUNNEL-01.
-//
-// # Blind-spot inventory (per ai-robust.md §"工具选定后强制盲区自检")
-//
-//   - Function-value reference `gen := bcrypt.GenerateFromPassword; gen(...)`:
-//     COVERED — the scan walks every <alias>.<sel> SelectorExpr (assignment RHS,
-//     arg pass, or CallExpr.Fun alike), not just call targets.
-//   - Dot-import `import . ".../bcrypt"; GenerateFromPassword(...)` (or the same
-//     for credential): the symbol becomes a bare *ast.Ident the SelectorExpr
-//     scan misses. The authoritative guard is the reverse self-test
-//     TestBCRYPT_COST_FUNNEL_01_NoDotImportBlindSpot, which asserts no file
-//     dot-imports either module (revive's dot-imports lint is a supplementary,
-//     not relied-upon, layer).
-//   - Reflection-based construction: out of scope, treated as theoretical.
 package archtest
 
 import (
 	"fmt"
-	"go/ast"
-	"go/parser"
-	"go/token"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -70,31 +12,6 @@ import (
 
 	"github.com/ghbvf/gocell/tools/archtest/internal/scanner"
 )
-
-const (
-	bcryptModulePath     = "golang.org/x/crypto/bcrypt"
-	credentialModulePath = "github.com/ghbvf/gocell/cells/accesscore/internal/credential"
-
-	// bcryptHasherRel is the single sanctioned file allowed to call
-	// bcrypt.GenerateFromPassword (A1). Module-relative, slash form.
-	bcryptHasherRel = "cells/accesscore/internal/credential/hasher.go"
-)
-
-// newTestHasherCallerAllowlist holds the module-relative locations permitted to
-// call credential.NewTestHasher (A2). *_test.go is handled separately by suffix;
-// this slice adds non-_test.go test-support packages (importable only by tests).
-//
-// accesscoretest is the sanctioned bridge letting external test packages (e.g.
-// tests/integration harnesses, which cannot import the internal credential
-// package under Go's internal rule) obtain a low-cost hasher via
-// accesscoretest.MinCostPasswordHasherOption(). That accesscoretest is imported ONLY
-// by tests is a convention, not a compiler-enforced barrier — this is the
-// upstream-Medium edge of the funnel (A1/A2 downstream are Hard). The
-// Hard-ization path is a generic "test-support packages imported only by
-// *_test.go" guard.
-var newTestHasherCallerAllowlist = []string{
-	"cells/accesscore/accesscoretest/", // test-support builders, imported only by *_test.go
-}
 
 // TestBCRYPT_COST_FUNNEL_01_A1_SingleHashOrigin fails if bcrypt.GenerateFromPassword
 // appears in any production (non-test) file other than credential/hasher.go.
@@ -154,64 +71,6 @@ func TestBCRYPT_COST_FUNNEL_01_A2_TestHasherCallerAllowlist(t *testing.T) {
 	assert.Empty(t, findings,
 		"credential.NewTestHasher is the low-cost test door; it may be called only from *_test.go "+
 			"or sanctioned test-support packages. Production wires credential.NewProductionHasher().")
-}
-
-func newTestHasherCallerAllowed(rel string) bool {
-	if strings.HasSuffix(rel, "_test.go") {
-		return true
-	}
-	for _, prefix := range newTestHasherCallerAllowlist {
-		if strings.HasPrefix(rel, prefix) {
-			return true
-		}
-	}
-	return false
-}
-
-// firstQualifiedSelectorLine returns the line of the first <alias>.<selName>
-// SelectorExpr in path, where <alias> is the local import name bound to
-// modulePath (handles named/aliased imports). A dot-import returns
-// ("",false) → not matched here; that blind spot is closed by the dedicated
-// reverse self-tests. Walking SelectorExpr (not just CallExpr.Fun) also catches
-// function-value references like `gen := pkg.Symbol`.
-func firstQualifiedSelectorLine(path, modulePath, defaultName, selName string) (int, bool, error) {
-	fset := token.NewFileSet()
-	file, err := parser.ParseFile(fset, path, nil, 0)
-	if err != nil {
-		return 0, false, err
-	}
-	alias, ok := moduleImportAlias(file, modulePath, defaultName)
-	if !ok {
-		return 0, false, nil
-	}
-	var pos token.Pos
-	scanner.EachInSubtree[ast.SelectorExpr](file, func(sel *ast.SelectorExpr) {
-		if pos.IsValid() {
-			return
-		}
-		id, isIdent := sel.X.(*ast.Ident)
-		if isIdent && id.Name == alias && sel.Sel.Name == selName {
-			pos = sel.Pos()
-		}
-	})
-	if pos.IsValid() {
-		return fset.Position(pos).Line, true, nil
-	}
-	return 0, false, nil
-}
-
-// moduleImportAlias returns the local import name bound to modulePath in file,
-// or ("",false) if not imported or dot/blank-imported.
-func moduleImportAlias(file *ast.File, modulePath, defaultName string) (string, bool) {
-	for _, imp := range file.Imports {
-		if archStringLiteralValue(imp.Path) != modulePath {
-			continue
-		}
-		if name := importSelectorName(imp, defaultName); name != "" {
-			return name, true
-		}
-	}
-	return "", false
 }
 
 // TestBCRYPT_COST_FUNNEL_01_A1_RedFixture asserts the A1 detector fires on a
@@ -282,21 +141,4 @@ func TestBCRYPT_COST_FUNNEL_01_NoDotImportBlindSpot(t *testing.T) {
 	}
 	assert.Empty(t, findings,
 		"dot-importing bcrypt or the credential package would make the funnel symbols bare idents and evade the scan; forbidden")
-}
-
-func fileDotImportsModule(path, modulePath string) (bool, error) {
-	fset := token.NewFileSet()
-	file, err := parser.ParseFile(fset, path, nil, 0)
-	if err != nil {
-		return false, err
-	}
-	for _, imp := range file.Imports {
-		if archStringLiteralValue(imp.Path) != modulePath {
-			continue
-		}
-		if imp.Name != nil && imp.Name.Name == "." {
-			return true, nil
-		}
-	}
-	return false, nil
 }
