+++ tools/archtest/no_deleted_auth_symbols_test.go
@@ -1,91 +1,8 @@
 // INVARIANT: NO-DELETED-AUTH-SYMBOLS-01
-//
-// # NO-DELETED-AUTH-SYMBOLS-01
-//
-// Invariant: no production or test .go file (in any package, including
-// runtime/auth/ itself) may reference or re-declare the deleted symbols
-//   - auth.RoleInternalAdmin
-//   - auth.ServiceNameInternal
-//   - auth.BuiltinServiceRoles
-//
-// These symbols were removed in Wave 2 of the SVCTOKEN-CALLER-IDENTITY
-// migration (PR #362 "A5 service token caller_cell + contract.clients
-// runtime enforce"). The rule enforces a literal "0 references / 0
-// re-declarations" check — re-introduction in ANY package (external,
-// dot-imported, OR runtime/auth itself) fails CI. There is NO package-level
-// carve-out for runtime/auth: the symbols were physically deleted, so
-// re-declaration or self-use inside runtime/auth is exactly what the rule
-// must catch (closing the BS-同包 gap raised in PR #1180 review).
-//
-// # AI-robust 评级：Medium-true (type-aware via typeseval.ResolvePackageRef)
-//
-// Resolution is by canonical *types.PkgName import path (info.Uses[sel.X]
-// .(*types.PkgName).Imported().Path()), NOT AST identifier name, so import
-// aliases (`import authz "...runtime/auth"; authz.RoleInternalAdmin`) cannot
-// bypass detection and same-name decoys (`import auth "other/pkg"`) cannot
-// trigger false positives. This closes the Soft-tier debt called out by
-// docs/reviews/202605181109-042-archtest-six-agent-audit.md §3a 合规红线第 2
-// 条; the upgrade follows ai-robust.md §Hard 范本目录 "string-typed concept
-// funnel" (Medium天花板 per 042 §3b: Go 类型系统对黑名单引用守护的客观上限).
-//
-// # Detection
-//
-// Three AST forms together cover every cross-package reference AND every
-// in-package re-declaration of a banned identifier:
-//   - (A) qualified SelectorExpr `pkg.X` — alias-transparent via
-//     info.Uses[sel.X].(*types.PkgName).Imported().Path() (covers consts +
-//     funcs at any expression position, including value-capture
-//     `var fn = auth.BuiltinServiceRoles`). Resolution delegates to
-//     ResolvePackageRef typed façade.
-//   - (B) bare *ast.Ident reference (info.Uses) — covers dot-imported
-//     references (`import . "...runtime/auth"; _ = RoleInternalAdmin / _ =
-//     ServiceNameInternal / BuiltinServiceRoles(...)`) AND same-package
-//     self-references inside authImportPath. Uses info.Uses[id] directly
-//     with type switch over {*types.Const, *types.Var, *types.Func,
-//     *types.TypeName}, bypassing ResolvePackageRef's typed-callable filter
-//     (which intentionally rejects Const/Var per its godoc).
-//   - (C) bare *ast.Ident declaration (info.Defs) — catches package-scope
-//     re-declaration inside authImportPath (`const RoleInternalAdmin =
-//     "..."` / `func BuiltinServiceRoles(...)`). Only fires when the pass
-//     IS authImportPath; declarations in other packages with matching
-//     names are unrelated symbols. Local scope (function bodies, struct
-//     fields) is excluded via obj.Parent() == obj.Pkg().Scope() check.
-//
-// # Blind spots
-//
-//   - BS-1 Reflection via reflect.Value.{Field,Method}ByName with constant
-//     string argument: NOT a Go identifier, so neither (A) nor (B) sees it.
-//     Reverse self-check TestNO_DELETED_AUTH_SYMBOLS_01_BS1_NoReflectAccess
-//     reuses the shared scanReflectStringArgCalls (REFLECT-STRING-ARG-SCANNER-01)
-//     to assert production AST has zero such references; companion
-//     TestNO_DELETED_AUTH_SYMBOLS_01_BS1_FixtureCatchesReflect pins the
-//     scanner-logic-verified contract via a fixture exercising both method-value
-//     and method-expression call forms. Covers chained shapes like
-//     `reflect.ValueOf(x).MethodByName(...)` and
-//     `reflect.Value.FieldByName(recv, ...)` uniformly.
-//   - BS-2 (accepted) `//go:linkname` directive: AST scanners do not parse
-//     compiler directives, so `//go:linkname myLocal
-//     github.com/.../runtime/auth.BuiltinServiceRoles` would not be detected
-//     by either (A) or (B). The symbols are physically deleted, so a
-//     linkname reference fails at link time (LINKLOAD does not resolve);
-//     this is a build-time guard rather than a CI archtest gap. Sole
-//     accepted residual.
-//
-// # Hard is unattainable for this rule shape
-//
-// "Deleted symbol reference 禁用" is a blacklist guard; Go's type system
-// has no per-symbol reference封禁 mechanism, and the Hard 范本目录
-// (sealing / single-sanctioned-holder / typed-marker / codegen-funnel) is
-// oriented at whitelist / single-construction-site semantics. The current
-// type-aware archtest is the Go ceiling per ai-robust.md §载体决策原则 ≥
-// Medium 立项硬门槛.
 package archtest
 
 import (
 	"fmt"
-	"go/ast"
-	"go/token"
-	"go/types"
 	"strings"
 	"testing"
 
@@ -93,202 +10,14 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-const ruleNoDeletedAuthSymbols01 = "NO-DELETED-AUTH-SYMBOLS-01"
-
-// fixtureAuthImportPath is the canonical path of the fixture-local fake
-// `auth` package used by TestNO_DELETED_AUTH_SYMBOLS_01_FixtureCatchesAllForms.
-// It is intentionally distinct from authRuntimeImportPath (defined in
-// svctoken_caller_cell_test.go) so the same scanner function exercises both
-// targets without redeclaring constants in production scope.
-const fixtureAuthImportPath = "github.com/ghbvf/gocell/tools/archtest/internal/nodeletedauthsymbolsfixture/auth"
-
-// deletedAuthSymbols is the set of selector / identifier names that must not
-// appear in any reference outside the canonical definition site.
-var deletedAuthSymbols = map[string]bool{
-	"RoleInternalAdmin":   true,
-	"ServiceNameInternal": true,
-	"BuiltinServiceRoles": true,
-}
-
-// scanDeletedAuthSymbolsAgainst walks a typed Pass and records every
-// reference OR package-scope re-declaration of a banned identifier whose
-// owning package resolves to authImportPath. Parametrized over authImportPath
-// so the production invariant test (authRuntimeImportPath) and the fixture
-// self-check (fixtureAuthImportPath) share one implementation.
-//
-// No package-level carve-out: the symbols were physically deleted in Wave 2,
-// so re-introduction inside authImportPath itself is exactly what the rule
-// must catch. The PR文案 "0 references / re-introduction fails CI" is
-// implemented literally — any reference or package-scope re-declaration in
-// any package (including authImportPath itself) is reported.
-func scanDeletedAuthSymbolsAgainst(p *Pass, authImportPath string) []Diagnostic {
-	if p.TypesInfo == nil {
-		return nil
-	}
-	selfPackage := p.Pkg != nil && p.Pkg.Path() == authImportPath
-	var diags []Diagnostic
-	for _, file := range p.Files {
-		rel := p.Rel(file)
-
-		// Pre-collect the positions of SelectorExpr.Sel so the bare-Ident
-		// scan below does not double-count an Ident that is already the Sel
-		// half of a qualified selector caught by (A). Without this guard,
-		// EachInSubtree[ast.Ident] would visit `auth.RoleInternalAdmin`'s
-		// `RoleInternalAdmin` Sel-Ident and resolve it (via info.Uses) to the
-		// same *types.Const / *types.Func that (A) already recorded — every
-		// qualified reference would be reported twice.
-		selSelPositions := make(map[token.Pos]bool)
-		EachInSubtree[ast.SelectorExpr](file, func(sel *ast.SelectorExpr) {
-			if sel.Sel != nil {
-				selSelPositions[sel.Sel.Pos()] = true
-			}
-		})
-
-		// (A) Qualified SelectorExpr — alias-transparent via ResolvePackageRef.
-		EachInSubtree[ast.SelectorExpr](file, func(sel *ast.SelectorExpr) {
-			pkgPath, name, ok := ResolvePackageRef(p.TypesInfo, sel)
-			if !ok || pkgPath != authImportPath {
-				return
-			}
-			if !deletedAuthSymbols[name] {
-				return
-			}
-			diags = append(diags, Diagnostic{
-				Rel:     rel,
-				Line:    p.Fset.Position(sel.Pos()).Line,
-				Message: formatBannedSymbolDiag(authImportPath, name, ""),
-			})
-		})
-
-		// (B) Bare *ast.Ident — Const / Var / Func / TypeName.
-		// Uses info.Uses[id] directly with type switch over the four
-		// package-level object kinds so both dot-imported bare refs AND
-		// same-package self-references inside authImportPath are detected
-		// (ResolvePackageRef intentionally rejects Const/Var per its
-		// typed-callable filter; that limitation is bypassed at caller site
-		// here, keeping the façade unchanged — façade extension is tracked
-		// separately by #1037).
-		EachInSubtree[ast.Ident](file, func(id *ast.Ident) {
-			if selSelPositions[id.Pos()] {
-				return
-			}
-			if !deletedAuthSymbols[id.Name] {
-				return
-			}
-			obj := p.TypesInfo.Uses[id]
-			if obj == nil || obj.Pkg() == nil || obj.Pkg().Path() != authImportPath {
-				return
-			}
-			switch obj.(type) {
-			case *types.Func, *types.Const, *types.Var, *types.TypeName:
-				// covered
-			default:
-				return
-			}
-			qualifier := " (dot-imported)"
-			if selfPackage {
-				qualifier = " (same-package self-reference)"
-			}
-			diags = append(diags, Diagnostic{
-				Rel:     rel,
-				Line:    p.Fset.Position(id.Pos()).Line,
-				Message: formatBannedSymbolDiag(authImportPath, obj.Name(), qualifier),
-			})
-		})
-
-		// (C) Package-scope re-declaration inside authImportPath. Catches
-		// `const RoleInternalAdmin = "..."` / `func BuiltinServiceRoles(...)`
-		// reintroductions that would otherwise sit in the package as
-		// dead-but-declared symbols (zero usage outside ⇒ (A)/(B) silent),
-		// breaking the "re-introduction fails CI" contract. Only runs when
-		// the pass IS the authImportPath; declarations in other packages
-		// with the same name are unrelated symbols.
-		if selfPackage {
-			EachInSubtree[ast.Ident](file, func(id *ast.Ident) {
-				if !deletedAuthSymbols[id.Name] {
-					return
-				}
-				obj := p.TypesInfo.Defs[id]
-				if obj == nil || obj.Pkg() == nil || obj.Pkg().Path() != authImportPath {
-					return
-				}
-				// Package scope only — local vars / params / struct fields
-				// that happen to share a name are not re-introductions of
-				// the deleted package-scope symbol.
-				if obj.Parent() == nil || obj.Parent() != obj.Pkg().Scope() {
-					return
-				}
-				switch obj.(type) {
-				case *types.Func, *types.Const, *types.Var, *types.TypeName:
-					// covered
-				default:
-					return
-				}
-				diags = append(diags, Diagnostic{
-					Rel:     rel,
-					Line:    p.Fset.Position(id.Pos()).Line,
-					Message: formatBannedSymbolDeclDiag(authImportPath, obj.Name()),
-				})
-			})
-		}
-	}
-	return diags
-}
-
-// formatBannedSymbolDiag composes the diagnostic message for a banned symbol
-// reference. The qualifier suffix differentiates the three reference shapes
-// — (A) qualified `pkg.X`, (B) dot-imported bare, (B) same-package self-ref
-// — so log readers can tell at a glance which AST form triggered.
-func formatBannedSymbolDiag(authImportPath, name, qualifier string) string {
-	return fmt.Sprintf(
-		"deprecated symbol %s.%s%s — replace with auth.RequireCallerCell (authz) "+
-			"or auth.TestServiceContext (test principals); see PR #362 SVCTOKEN-CALLER-IDENTITY",
-		shortPkg(authImportPath), name, qualifier,
-	)
-}
-
-// formatBannedSymbolDeclDiag composes the diagnostic message for a
-// package-scope re-declaration of a deleted symbol inside authImportPath.
-// Distinct from formatBannedSymbolDiag because the actionable advice differs
-// (delete the re-introduced declaration; do not just rewrite call sites).
-func formatBannedSymbolDeclDiag(authImportPath, name string) string {
-	return fmt.Sprintf(
-		"deleted symbol %s.%s re-declared at package scope — delete the declaration; "+
-			"these symbols were removed in Wave 2 of SVCTOKEN-CALLER-IDENTITY (PR #362) "+
-			"and callers must migrate to auth.RequireCallerCell (authz) or "+
-			"auth.TestServiceContext (test principals)",
-		shortPkg(authImportPath), name,
-	)
-}
-
-// productionScanPatterns is the seven production roots scanned by both the
-// main invariant test and the BS-1 reverse self-check.
-var productionScanPatterns = []string{
-	"./runtime/...",
-	"./cells/...",
-	"./cmd/...",
-	"./kernel/...",
-	"./adapters/...",
-	"./examples/...",
-	"./tests/...",
-}
-
 // TestNO_DELETED_AUTH_SYMBOLS_01 enforces that no production or test code
 // references the three deleted runtime/auth symbols. The scan runs typed
 // over the seven production roots, paired with FlatNonDefaultTags() to cover
 // build-tagged variants (e.g. //go:build integration test helpers).
 func TestNO_DELETED_AUTH_SYMBOLS_01(t *testing.T) {
 	t.Parallel()
-
-	diags := Run(t, Typed(
-		TypedOpts{Tests: true, Tags: FlatNonDefaultTags()},
-		productionScanPatterns,
-	),
-		func(p *Pass) []Diagnostic {
-			return scanDeletedAuthSymbolsAgainst(p, authRuntimeImportPath)
-		})
-
-	Report(t, ruleNoDeletedAuthSymbols01, diags)
+	Report(t, ruleNoDeletedAuthSymbols01,
+		CheckNoDeletedAuthSymbols01(t, ConfigForExternalCell{BuildTags: FlatNonDefaultTags()}))
 }
 
 // TestNO_DELETED_AUTH_SYMBOLS_01_FixtureCatchesAllForms exercises every
