+++ tools/archtest/testdata/module_path_funnel.baseline
@@ -9,20 +9,14 @@
 tools/archtest/adapter_returns_declared_types_test.go
 tools/archtest/archtest_test.go
 tools/archtest/audit_hash_input_frozen_test.go
-tools/archtest/auth_authtest_boundary_test.go
-tools/archtest/auth_keystest_boundary_test.go
-tools/archtest/auth_plan_test.go
-tools/archtest/bcrypt_cost_funnel_test.go
 tools/archtest/capability_provider_funnel_test.go
 tools/archtest/celltest_import_boundary_test.go
 tools/archtest/celltest_import_scope_test.go
 tools/archtest/clock_invariants_test.go
 tools/archtest/contract_path_query_coverage_test.go
 tools/archtest/contracttest_loadbyid_literal_test.go
-tools/archtest/ctxkeys_principal_write_caller_test.go
 tools/archtest/distlock_lock_not_context_test.go
 tools/archtest/eval_predicate_centralization_test.go
-tools/archtest/fence_token_mint_funnel_test.go
 tools/archtest/fixture_cellid_typed_builder_test.go
 tools/archtest/governance_rules_invariants_test.go
 tools/archtest/health_aggregation_test.go
@@ -33,7 +27,6 @@ tools/archtest/l2_outbox_atomicity_coverage_test.go
 tools/archtest/metrics_cancel_ctx_conformance_test.go
 tools/archtest/mqtt_funnel_test.go
 tools/archtest/mqtt_reason_redaction_test.go
-tools/archtest/no_deleted_auth_symbols_test.go
 tools/archtest/notfound_test_strict_test.go
 tools/archtest/observability_metrics_test.go
 tools/archtest/outbox_invariants_test.go
@@ -52,12 +45,9 @@ tools/archtest/relay_isolation_test.go
 tools/archtest/required_dep_nil_guard_test.go
 tools/archtest/reverse_coverage_invariants_test.go
 tools/archtest/rmq_invariants_test.go
-tools/archtest/safeid_funnel_test.go
 tools/archtest/saga_invariants_test.go
 tools/archtest/security_defaults_test.go
-tools/archtest/seed_role_iface_test.go
 tools/archtest/serviceowned_handler_owner_check_test.go
-tools/archtest/svctoken_caller_cell_test.go
 tools/archtest/test_polling_external_reason_literal_test.go
 tools/archtest/webhook_hmac_funnel_test.go
 tools/archtest/wrapper_location_test.go
